backend/requirements.txt
psycopg2==2.8.5
django-nested-admin==4.0.2
redis==4.4.2
celery==5.2.7
requests==2.28.2
beautifulsoup4==4.11.1

asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
Markdown==3.4.1
pytz==2022.4
sqlparse==0.4.3

